requirements.txt
python-telegram-bot[job-queue]==20.8
python-dotenv==1.0.1
pandas==2.2.2
openai==1.35.7
httpx[socks,http2]==0.27.0
beautifulsoup4==4.12.3
feedparser==6.0.11
requests==2.32.3
matplotlib==3.9.0
